% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{closest_dist}
\alias{closest_dist}
\title{Closest distances between counties and a storm track}
\format{An object of class \code{data.frame} with 876936 rows and 4 columns.}
\usage{
closest_dist
}
\description{
Closest distances between counties and a storm track
}
\keyword{datasets}

