% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_exposure.R
\name{county_distance}
\alias{county_distance}
\title{Hurricane exposure by distance for counties}
\usage{
county_distance(counties, start_year, end_year, dist_limit)
}
\arguments{
\item{counties}{Character string of the five-digit FIPS codes for
counties for which the user wants to create data}

\item{start_year}{Four-digit integer with first year to consider.}

\item{end_year}{Four-digit integer with last year to consider.}

\item{dist_limit}{Maximum distance, in kilometers, for the closest distance
between the county center and the storm track to consider
the county "exposed" to the storm.}
}
\value{
Returns a dataframe with a row for each county-storm
   pair and with columns giving the storm's ID, county FIPS,
   date and time of closest approach of that storm to that county, and
   distance between the county center and the storm track on
   that date and time.
}
\description{
This function takes a list of county FIPS, boundaries on
starting and ending years, and thresholds for distance
and returns a dataframe with the subset of storms meeting those criteria
for each of the listed counties.
}
\examples{
county_distance(counties = c("22071", "51700"),
            start_year = 1995, end_year = 2005,
            dist_limit = 75)

}

