% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_exposure.R
\name{county_events}
\alias{county_events}
\title{Find events for storms by county}
\usage{
county_events(counties, start_year, end_year, event_type)
}
\description{
Find events for storms by county
}
\examples{
county_events(counties = c("37031", "37053", "37055"),
              start_year = 1999, end_year = 1999,
              event_type = "flood")
}

