% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind_exposure.R
\name{county_wind}
\alias{county_wind}
\title{Hurricane exposure by wind for counties}
\usage{
county_wind(counties, start_year, end_year, wind_limit,
  wind_var = "vmax_sust")
}
\arguments{
\item{counties}{Character vector of the five-digit Federal Information
Processing Standard (FIPS) codes for counties for which the user wants to
determine storm exposure.}

\item{start_year}{Four-digit integer with first year to consider.}

\item{end_year}{Four-digit integer with last year to consider.}

\item{wind_limit}{A numeric vector of length one giving the minimum
wind speed (in meters per second) to use in the filter}

\item{wind_var}{A character string giving the wind variable to use. Choices
are "vmax_sust" (maximum sustained winds; default) or "vmax_gust" (maximum
gust winds).}
}
\value{
Returns a dataframe with a row for each county-storm
   pair and with columns for:
   \itemize{
     \item{\code{storm_id}: }{Unique storm identifier with the storm name and year,
                 separated by a hyphen(e.g., "Alberto-1988",
                 "Katrina-2005")}
     \item{\code{fips}: }{County's 5-digit Federal Information Processing Standard
                 (FIPS) code}
     \item{\code{max_sust}: }{Maximum sustained wind speed (in m / s)}
     \item{\code{max_gust}: }{Maximum gust wind speed (in m / s)}
   }
}
\description{
This function takes a list of US counties,based on their 5-digit Federal
Information Processing Standard (FIPS) codes, boundaries on
the range of years to be considered, and thresholds for wind speed (in meters
per second) for each  county to be considered "exposed" to the
storm. Based on these inputs, the function returns a dataframe with the
subset of Atlantic basin storms meeting those criteria for each of the listed
counties.
}
\details{
For more information on how wind speeds are modeled in this data,
   see the documentation for the \code{stormwindmodel} R package.
}
\note{
Only counties in states in the eastern half of the United States can
   be processed by this function.
}
\examples{
county_wind(counties = c("22071", "51700"),
            start_year = 1988, end_year = 2005,
            wind_limit = 20, wind_var = "vmax_sust")

}

