% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_exposure.R
\name{default_map}
\alias{default_map}
\title{Create a default map with eastern US states}
\usage{
default_map()
}
\value{
A ggplot object that maps the states of the Eastern United States
}
\description{
Creates a ggplot object with the underlying map of all states in the
eastern section of the US that might be prone to hurricane-related
exposure. Other lines and points can be added to the output using
`ggplot2` plotting functions.
}
\examples{
default_map()
}

