% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_exposure.R
\name{distance_exposure}
\alias{distance_exposure}
\title{Create storm exposure time series files}
\usage{
distance_exposure(locations, start_year, end_year, dist_limit, out_dir,
  out_type = "csv")
}
\arguments{
\item{locations}{Either a vector of FIPS county codes, for county-level
output, or a dataframe with columns for community identifier (\code{commun})
and associated FIPS codes (\code{fips}), for multi-county community output}

\item{start_year}{Four-digit integer with first year to consider.}

\item{end_year}{Four-digit integer with last year to consider.}

\item{dist_limit}{Maximum distance, in kilometers, for the closest distance
between the county center and the storm track to consider
the county "exposed" to the storm.}

\item{out_dir}{Character string giving the pathname of the directory in which
to write output. This directory should already exist on your computer.}

\item{out_type}{Character string giving the type of output files you'd like.
Options are \code{"csv"} (default) and \code{"rds"}.}
}
\value{
Writes out a directory with rain exposure files for each county or
   community indicated. For more on the columns in this output, see the
   documentation for \code{\link{county_rain}} and
   \code{\link{multi_county_rain}}.
}
\description{
This function takes an input of locations (either a vector of county FIPS
or a dataframe of multi-county FIPS, with all FIPS listed for each county)
and creates time series dataframes that can be merged with health time series,
giving the dates and exposures for all storms meeting the given rainfall and
storm distance criteria.
}
\examples{
# By county
distance_exposure(locations = c("22071", "51700"),
              start_year = 1995, end_year = 2005,
              dist_limit = 25,
              out_dir = "~/tmp/storms")

# For multi-county communities
communities <- data.frame(commun = c(rep("ny", 6), "no", "new"),
                          fips = c("36005", "36047", "36061",
                          "36085", "36081", "36119",
                          "22071", "51700"))
distance_exposure(locations = communities,
              start_year = 1995, end_year = 2005,
              dist_limit = 25,
              out_dir = "~/tmp/storms")

}

