% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_storm_data.R
\name{filter_wind_data}
\alias{filter_wind_data}
\title{Filter hurricane wind dataset}
\usage{
filter_wind_data(counties = NULL, storm = NULL, year_range = NULL,
  wind_limit = NULL, output_vars = "fips", wind_var = "vmax_sust")
}
\arguments{
\item{counties}{A character vector listing all 5-digit county FIPS codes
for a subset of counties.}

\item{storm}{A character string giving the storm ID (e.g., "Floyd-1999")}

\item{year_range}{A numeric vector of length two with the starting and
ending year to subset to.}

\item{wind_limit}{A numeric vector of length one giving the minimum
wind speed (in meters per second) to use in the filter}

\item{output_vars}{A character vector listing all the columns to include
in the output.}

\item{wind_var}{A character string giving the wind variable to use. Choices
are "vmax_sust" (maximum sustained winds; default) or "vmax_gust" (maximum
gust winds).}
}
\value{
A dataframe with storms filtered based on the input criteria to the
   function. Columns in the output will vary depending on the user's
   selections for the \code{output_vars} argument.
}
\description{
This function is a helper function for many of the the other functions in
this package that measure wind exposure.
}
\examples{
filter_wind_data(counties = c("22071", "51700"), year_range = c(1988, 2011),
                 wind_limit = 20,
                 output_vars = c("fips", "storm_id", "vmax_sust"))

}

