% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_exposure.R
\name{get_eastern_map}
\alias{get_eastern_map}
\title{Get map data for eastern US states}
\usage{
get_eastern_map(map = "county")
}
\arguments{
\item{map}{A character string giving the map database from which to pull}
}
\value{
A dataframe with map data pulled using the \code{map_data} function in
   \code{ggplot2}, filtered to states in the eastern half of the United States.
}
\description{
Get map data for eastern US states
}
