% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_exposure.R
\name{hurr_choropleth}
\alias{hurr_choropleth}
\title{Create a map customized for this package}
\usage{
hurr_choropleth(map_data, metric = "distance")
}
\arguments{
\item{map_data}{A dataframe with columns with FIPS numbers (in numeric
class) for all counties in the eastern US (\code{region}) and the
exposure value (\code{value})}

\item{metric}{Character string giving the metric to plot. Current options are
\code{"distance"} (default) and \code{"rainfall"}. These options are used
to customize the color palette and scale of the choropleth map produced
by this function.}
}
\value{
A \code{ggplot} object with a map of hurricane exposure in eastern
   US counties
}
\description{
Creates a county choropleth map customized for displaying
hurricane exposure.
}
\details{
The function only maps counties in states likely to be exposed
to Atlantic basin tropical storms.
}

