% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_exposure.R
\name{hurr_choroplethr}
\alias{hurr_choroplethr}
\title{Create a map customized for this package}
\usage{
hurr_choroplethr(map_data, metric = "distance")
}
\arguments{
\item{map_data}{A dataframe with columns with FIPS numbers (\code{region})
and the exposure value (\code{value})}

\item{metric}{Character string giving the metric to plot. Current options are
\code{"distance"} (default) and \code{"rainfall"}.}
}
\value{
A \code{choroplethr} object. To plot the map, use the \code{render}
   method.
}
\description{
Create a map customized for this package
}

