% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hurr_tracks}
\alias{hurr_tracks}
\title{Storm tracks for Atlantic basin storms}
\format{A data frame with 12,206 rows and 6 variables:
\describe{
  \item{date}{Date and time of storm track recording}
  \item{status}{Storm status}
  \item{latitude}{Latitude of storm center}
  \item{longitude}{Longitude of storm center}
  \item{wind}{[what is this measuring?]}
  \item{storm_id}{Unique storm identifier}
}}
\usage{
hurr_tracks
}
\description{
A dataset containing the storm tracks for all Atlantic basin tropical
storms between 1987 and 2012.
}
\note{
The original dataset included more separate `status` categories
   than this version. Here, I have pooled together into "Other" the
   following categories: extratropical cyclones (of any intensity),
   subtropical cycles of both subtropical depression and subtropical
   storm intensity, tropical waves of any intensity, disturbances of
   any intensity, and lows that do not fall under any of these
   categories.
}
\references{
[Determine correct refenence]
}
\keyword{datasets}

