% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_exposure.R
\name{interp_track}
\alias{interp_track}
\title{Interpolate a storm track}
\usage{
interp_track(track)
}
\arguments{
\item{track}{A dataframe with hurricane track data for a single storm}
}
\value{
A dataframe with hurricane track data for a single storm,
   interpolated to 15-minute intervals.
}
\description{
This function takes a wider-spaced storm track (e.g., every 6 hours) and
interpolates to every 15 minutes. To do this, it
fits GLMs of latitude and longitude regressed on natural cubic splines of
date-time, and then predicts these splines to new intervals. These
splines use degrees of freedom equal to the number of original observations
divided by two.
}
