% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_exposure.R
\name{map_counties}
\alias{map_counties}
\title{Map counties}
\usage{
map_counties(storm, metric = "distance", days_included = c(-2, -1, 0, 1))
}
\arguments{
\item{storm}{Character string giving the name of the storm to plot (e.g.,
"Floyd-1999")}

\item{metric}{Character string giving the metric to plot. Current options are
\code{"distance"} (default) and \code{"rainfall"}. These options are used
to customize the color palette and scale of the choropleth map produced
by this function.}

\item{days_included}{A numeric vector listing the days to include when
calculating total precipitation. Negative numbers are days before the
closest date of the storm to a county. For example,
\code{c(-1, 0, 1)} would calculate rain for a county as the sum of the
rainfall for the day before, the day of, and the day after the date when
the storm center was closest to the county center. Values can range from
-3 to 3 (i.e., at most, you can calculate the total rainfall for the
seven-day window centered on the day when the storm is closest to the
county).}
}
\value{
This function creates a choropleth map of counties in the eastern
   part of the United States, showing distance from a storm track or total
   rainfall over a given window of one or more days.
}
\description{
Map counties
}
\examples{
\dontrun{
floyd_map <- map_counties("Floyd-1999", metric = "rainfall",
                           days_included = c(-2, -1, 0, 1))
floyd_map

beryl_map <- map_counties("Beryl-1988", metric = "wind")
}
}

