% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_exposure.R
\name{map_distance_exposure}
\alias{map_distance_exposure}
\title{Map counties with distance exposure}
\usage{
map_distance_exposure(storm, dist_limit)
}
\arguments{
\item{storm}{Character string giving the name of the storm to plot (e.g.,
"Floyd-1999")}
}
\value{
Plots a map showing whether eastern US counties were exposed or
   unexposed to a specific storm based on a distance criterion.
}
\description{
Map counties as "exposed" or "unexposed" based on the criterion that the
storm came within a given distance (specified by \code{dist_limit}) of
the county's population mean center.
}
\examples{

floyd_map <- map_distance_exposure(storm = "Floyd-1999", dist_limit = 75)
floyd_map

allison_map <- map_distance_exposure(storm = "Allison-2001",
                                     dist_limit = 75)
map_tracks("Allison-2001", plot_points = FALSE, plot_object = allison_map)

}

