% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_exposure.R
\name{map_event_exposure}
\alias{map_event_exposure}
\title{Map county-level exposure based on reported events}
\usage{
map_event_exposure(storm_id, event_type)
}
\description{
Map county-level exposure based on reported events
}
\examples{
map_event_exposure(storm_id = "Floyd-1999", event_type = "flood")
map_event_exposure(storm_id = "Floyd-1999", event_type = "tornado")
map_event_exposure(storm_id = "Floyd-1999", event_type = "wind")
map_event_exposure(storm_id = "Floyd-1999", event_type = "tropical_storm")

}

