% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_exposure.R
\name{map_wind_exposure}
\alias{map_wind_exposure}
\title{Map counties with wind exposure}
\usage{
map_wind_exposure(storm, wind_limit)
}
\arguments{
\item{storm}{A character string giving the storm ID (e.g., "Floyd-1999")}

\item{wind_limit}{A numeric vector of length one giving the minimum
wind speed (in meters per second) to use in the filter}
}
\value{
Plots a map showing whether eastern US counties were exposed or
   unexposed to a specific storm based on a wind criterion.
}
\description{
Map counties as "exposed" or "unexposed" based on the criterion that the
storm had a certain wind speed (specified by \code{wind_limit}, in meters
per second).
}
\examples{

beryl_map <- map_wind_exposure(storm = "Beryl-1988", wind_limit = 15)
beryl_map
map_tracks("Beryl-1988", plot_points = FALSE, plot_object = beryl_map)

}

