% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_exposure.R
\name{multi_county}
\alias{multi_county}
\title{Create multi-county community exposures}
\usage{
multi_county(rain_storm_df, communities)
}
\arguments{
\item{rain_storm_df}{A dataframe of county-level "exposed"
storms, as created by \code{rain_storms}.}

\item{community_list}{A list object with an element for each
community and a vector of FIPS codes for all counties within
each community}
}
\value{
Returns the same type dataframe as \code{rain_storms},
   but with storms listed by community instead of county.
}
\description{
This function takes a rain-exposure dataframe for counties,
   as created by the \code{rain_storms} function, as well as
   a list object with the counties in each multi-county
   community, and returns a community-level dataframe of
   "exposed" storms, based on the rainfall and distance
   thresholds used to create the rain-exposure dataframe.
}
\examples{
communities <- data.frame(commun = c(rep("ny", 6), "no", "new"),
                         fips = c("36005", "36047", "36061",
                                  "36085", "36081", "36119",
                                  "22071", "51700"))
rain_storm_df <- rain_storms(counties = communities$fips,
                             start_year = 1995, end_year = 2005,
                             rain_limit = 20, dist_limit = 100)
}

