% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{precip_file}
\alias{precip_file}
\title{Weekly rainfall totals for tropical storms}
\format{A list with 991,530 rows and 3 variables:
\describe{
  \item{storm_id}{unique identifier for each tropical storm}
  \item{fips}{county 5-digit FIPS code}
  \item{tot_precip}{Total precipitation, in millimeters [?], for a
      one-week window centered at the date the storm was closest to
      the county}
}}
\usage{
precip_file
}
\description{
A dataframe that gives the total rainfall in US counties for a one-week
window centered at the date on which the tropical storm was closest
to the county for all Atlantic basin storms between 1987 and 2012.
}
\references{
[Get proper reference from Bill and Mohammad]
}
\keyword{datasets}

