% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_exposure.R
\name{rain_storms}
\alias{rain_storms}
\title{Create subset of storms for FIPS}
\usage{
rain_storms(counties, start_year, end_year, rain_limit, dist_limit)
}
\arguments{
\item{counties}{Character string of the FIPS codes for
counties to create data for}

\item{start_year}{Four-digit integer with first year to consider.}

\item{end_year}{Four-digit integer with last year to consider.}

\item{rain_limit}{Minimum of rainfall for the week-long period
centered at the storm's closest date to consider county
"exposed" to the storm.}

\item{dist_limit}{Maximum distance for the closest distance
between the county center and the storm track to consider
the county "exposed" to the storm.}
}
\value{
Returns a dataframe with a row for each county-storm
   pair and with columns giving the storm's ID, county FIPS,
   date and time of closest approach of that storm to that county,
   distance between the county center and the storm track on
   that date and time, and cumulative rainfall for the week-long
   period centered at the date of closest approach to the county.
}
\description{
This function takes a list of county FIPS, boundaries on
   starting and ending years, and thresholds for distance
   and the week-window rainfall and returns a dataframe
   with the subset of storms meeting those criteria for
   each of the listed counties.
}
\examples{
rain_storms(counties = c("22071", "51700"),
            start_year = 1995, end_year = 2005,
            rain_limit = 100, dist_limit = 100)

}

