% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{storm_rains}
\alias{storm_rains}
\title{Weekly rainfall totals for tropical storms}
\format{A list with 991,530 rows and 3 variables:
\describe{
  \item{storm_id}{unique identifier for each tropical storm}
  \item{fips}{county 5-digit FIPS code}
  \item{tot_precip}{Total precipitation, in [unit], for a one-week
      window centered at the date the storm was closest to the county}
}}
\usage{
storm_rains
}
\description{
A list that gives the total rainfall in US counties for a one-week
window centered at the date on which the tropical storm was closest
to the county.
}
\keyword{datasets}

