% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind_exposure.R
\name{wind_exposure}
\alias{wind_exposure}
\title{Write storm wind exposure files}
\usage{
wind_exposure(locations, start_year, end_year, wind_limit, out_dir,
  out_type = "csv")
}
\arguments{
\item{locations}{Either a vector of FIPS county codes, for county-level
output, or a dataframe with columns for community identifier (\code{commun})
and associated FIPS codes (\code{fips}), for multi-county community output.
See the examples for the proper format for this argument.}

\item{start_year}{Four-digit integer with first year to consider.}

\item{end_year}{Four-digit integer with last year to consider.}

\item{wind_limit}{Vector giving the wind speed (in m / s) to use as a
threshold for classifying a county as "exposed" to a specific storm.}

\item{out_dir}{Character string giving the pathname of the directory in which
to write output. This directory should already exist on your computer.}

\item{out_type}{Character string giving the type of output files you'd like.
Options are \code{"csv"} (default) and \code{"rds"}.}
}
\value{
Writes out a directory with rain exposure files for each county or
   community indicated. For more on the columns in this output, see the
   documentation for \code{\link{county_wind}} and
   \code{\link{multi_county_wind}}.
}
\description{
This function takes an input of locations (either a vector of county FIPS
or a dataframe of multi-county FIPS, with all FIPS listed for each county)
and creates time series dataframes that can be merged with health time series,
giving the dates and exposures for all storms meeting the given
storm wind criteria.
}

