% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{county_centers}
\alias{county_centers}
\title{Location of United States county centers of population}
\format{A data frame with 3221 rows and 8 variables:
\describe{
  \item{fips}{County's 5-digit Federal Information Processing Standard (FIPS)
             code}
  \item{county_name}{County name}
  \item{state_name}{State name}
  \item{population}{Population of the county as of the 2010 US Census}
  \item{latitude}{Latitude of county's center of population, in decimal
                 degrees}
  \item{longitude}{Longitude of county's center of population, in decimal
                  degrees (note: longitudes are given as negative values for
                  longitudes west of the prime meridian)}
}}
\source{
\url{https://www.census.gov/geo/reference/centersofpop.html}
}
\usage{
county_centers
}
\description{
A dataset containing the latitude and longitude of the
center of population of every United State county,
based on population as of the 2010 US Census.
}
\details{
The latitude and longitude of each county are the county's
population-weighted mean center of population, based on population as of the
2010 US Census. For more details on the calculation of these mean centers of
population, see the reference below.
}
\references{
Bureau of the Census, Centers of Population Computation for the United
States 1950-2010, U.S. Department of Commerce, Bureau of the Census,
  Washington, DC, issued 2011. \url{http://www2.census.gov/geo/pdfs/reference/cenpop2010/COP2010_documentation.pdf}
}
\keyword{datasets}

