% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_exposure.R
\name{county_events}
\alias{county_events}
\title{Find events for storms by county}
\usage{
county_events(counties, start_year, end_year, event_type)
}
\arguments{
\item{counties}{Character vector of the five-digit Federal Information
Processing Standard (FIPS) codes for counties for which the user wants to
determine storm exposure.}

\item{start_year}{Four-digit integer with first year to consider.}

\item{end_year}{Four-digit integer with last year to consider.}

\item{event_type}{Character string with the type of event to use to identify
county exposures. Options include \code{"flood"}, \code{"tornado"},
\code{"wind"}, and \code{"tropical storm"}.}
}
\description{
Find events for storms by county
}
\note{
Of the event types that this function can pull, only "tornado" and
   "wind" are available for years before 1996. Therefore, this function will
   pull listed tornado events or wind events for all years, but for any other
   event type, you should not use a start year prior to 1996, as events of
   other types were not recorded in the database before 1996.
}
\examples{
\dontrun{
county_events(counties = c("37031", "37053", "37055"),
              start_year = 1988, end_year = 1999,
              event_type = "flood")
}
}

