% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hurr_tracks}
\alias{hurr_tracks}
\title{Storm tracks for Atlantic basin storms}
\format{An object of class \code{data.frame} with 11829 rows and 5 columns.}
\source{
\url{http://rammb.cira.colostate.edu/research/tropical_cyclones/tc_extended_best_track_dataset}
}
\usage{
hurr_tracks
}
\description{
A dataset containing the storm tracks for Atlantic basin tropical
storms between 1988 and 2012, from the Extended Best Track Dataset for
the Atlantic basin.
}
\details{
This extended best tracks dataset is currently available for
1988 to 2014. This dataset is based on the HURDAT best tracks dataset.
For more details, see the source and reference given below.
}
\references{
Demuth J, DeMaria M, Knaff JA, 2006. Improvement of advanced microwave
sounder unit tropical cyclone intensity and size estimation algorithms.
Journal of Applied Meteorology 45:1573-1581
}
\keyword{datasets}

