% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_exposure.R
\name{interp_track}
\alias{interp_track}
\title{Interpolate a storm track}
\usage{
interp_track(track)
}
\arguments{
\item{track}{A dataframe with hurricane track data for a single storm}
}
\value{
A dataframe with hurricane track data for a single storm,
   interpolated to 15-minute intervals.
}
\description{
This function takes a wider-spaced storm track (e.g., every 6 hours) and
interpolates to every 15 minutes. To do this, it uses natural cubic
spline interpolation using the `spline` function from the `stats` package.
The track is only interpolated if there are three or more observations
on the central location of the storm (this is almost always the case
for storms tracked in the HURDAT2 dataset).
}
