% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_exposure.R
\name{map_counties}
\alias{map_counties}
\title{Map counties}
\usage{
map_counties(storm, metric = "distance", days_included = c(-1, 0, 1))
}
\arguments{
\item{storm}{Character string giving the name of the storm to plot (e.g.,
"Floyd-1999")}

\item{metric}{Character string giving the metric to plot. Current options are
\code{"closest distance"} (default) and \code{"rainfall"}.}

\item{days_included}{A numeric vector listing the days to include when
calculating total precipitation. Negative numbers are days before the
closest date of the storm to a county. For example,
\code{c(-1, 0, 1)} would calculate rain for a county as the sum of the
rainfall for the day before, the day of, and the day after the date when
the storm center was closest to the county center.}
}
\description{
Map counties
}
\examples{
floyd_map <- map_counties("Floyd-1999", metric = "rainfall",
                           days_included = c(-1, 0, 1))
plot(floyd_map)

}

