% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_exposure.R
\name{map_distance_exposure}
\alias{map_distance_exposure}
\title{Map counties with distance exposure}
\usage{
map_distance_exposure(storm, dist_limit)
}
\arguments{
\item{dist_limit}{Maximum distance, in kilometers, for the closest distance
between the county center and the storm track to consider
the county "exposed" to the storm.}
}
\description{
Map counties with distance exposure
}
\examples{

floyd_map <- map_distance_exposure(storm = "Floyd-1999", dist_limit = 75)
plot(floyd_map)

allison_map <- map_distance_exposure(storm = "Allison-2001",
                                     dist_limit = 75)
map_tracks("Allison-2001", plot_points = FALSE, plot_object = allison_map,
           storm_status = FALSE)

}

