% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_exposure.R
\name{map_event_exposure}
\alias{map_event_exposure}
\title{Map county-level exposure based on reported events}
\usage{
map_event_exposure(storm_id, event_type)
}
\arguments{
\item{storm_id}{Character vector with the storm for which to map events
(e.g., \code{"Katrina-2005"})}

\item{event_type}{Character string with the type of event to use to identify
county exposures. Options include \code{"flood"}, \code{"tornado"},
\code{"wind"}, and \code{"tropical storm"}.}
}
\description{
Map county-level exposure based on reported events
}
\examples{
map_event_exposure(storm_id = "Floyd-1999", event_type = "flood")
map_event_exposure(storm_id = "Floyd-1999", event_type = "tornado")
map_event_exposure(storm_id = "Floyd-1999", event_type = "wind")
map_event_exposure(storm_id = "Floyd-1999", event_type = "tropical_storm")

}

