% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{map_rain_exposure}
\alias{map_rain_exposure}
\title{Map counties with rain exposure}
\usage{
map_rain_exposure(storm, rain_limit, dist_limit, days_included = c(-1, 0, 1))
}
\arguments{
\item{rain_limit}{Minimum of rainfall, in millimeters, summed across the days
selected to be included, that must be met to consider county
"exposed" to the storm.}

\item{dist_limit}{Maximum distance, in kilometers, for the closest distance
between the county center and the storm track to consider
the county "exposed" to the storm.}

\item{days_included}{A numeric vector listing the days to include when
calculating total precipitation. Negative numbers are days before the
closest date of the storm to a county. For example,
\code{c(-1, 0, 1)} would calculate rain for a county as the sum of the
rainfall for the day before, the day of, and the day after the date when
the storm center was closest to the county center.}
}
\description{
Map counties with rain exposure
}
\examples{

floyd_map <- map_rain_exposure(storm = "Floyd-1999", rain_limit = 50,
                               dist_limit = 100)
plot(floyd_map)

allison_map <- map_rain_exposure(storm = "Allison-2001", rain_limit = 20,
                                 dist_limit = 100, days_included = 0)
map_tracks("Allison-2001", plot_points = FALSE, plot_object = floyd_map,
           storm_status = FALSE)

}

