% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_exposure.R
\name{multi_county_rain}
\alias{multi_county_rain}
\title{Create multi-county community exposures}
\usage{
multi_county_rain(communities, start_year, end_year, rain_limit, dist_limit)
}
\arguments{
\item{communities}{A dataframe with the FIPS codes for all counties within
each community}

\item{start_year}{Four-digit integer with first year to consider.}

\item{end_year}{Four-digit integer with last year to consider.}

\item{rain_limit}{Minimum of rainfall for the week-long period
centered at the storm's closest date to consider county
"exposed" to the storm.}

\item{dist_limit}{Maximum distance for the closest distance
between the county center and the storm track to consider
the county "exposed" to the storm.}
}
\value{
Returns the same type dataframe as \code{rain_storms},
   but with storms listed by community instead of county.
}
\description{
This function takes a rain-exposure dataframe for counties,
   as created by the \code{rain_storms} function, as well as
   a list object with the counties in each multi-county
   community, and returns a community-level dataframe of
   "exposed" storms, based on the rainfall and distance
   thresholds used to create the rain-exposure dataframe.
}
\examples{
communities <- data.frame(commun = c(rep("ny", 6), "no", "new"),
                         fips = c("36005", "36047", "36061",
                                  "36085", "36081", "36119",
                                  "22071", "51700"))
rain_storm_df <- multi_county_rain(communities = communities,
                                   start_year = 1995, end_year = 2005,
                                   rain_limit = 100, dist_limit = 100)

}

