% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rain_exposure.R
\name{multi_county_rain}
\alias{multi_county_rain}
\title{Hurricane exposure by rain for communities}
\usage{
multi_county_rain(communities, start_year, end_year, rain_limit, dist_limit,
  days_included = c(-1, 0, 1))
}
\arguments{
\item{communities}{A dataframe with the FIPS codes for all counties within
each community. It must include columns with a column identifier
(\code{commun}) and with the FIPS codes of counties included in each
community (\code{fips}). See the example code.}

\item{start_year}{Four-digit integer with first year to consider.}

\item{end_year}{Four-digit integer with last year to consider.}

\item{rain_limit}{Minimum of rainfall, in millimeters, summed across the days
selected to be included (\code{days_included}), that must fall in a
county for the county to be classified as "exposed" to the storm.}

\item{dist_limit}{Maximum distance, in kilometers, of how close the storm
track must come to the county's population mean center to classify
the county as "exposed" to the storm.}

\item{days_included}{A numeric vector listing the days to include when
calculating total precipitation. Negative numbers are days before the
closest date of the storm to a county. For example,
\code{c(-1, 0, 1)} would calculate rain for a county as the sum of the
rainfall for the day before, the day of, and the day after the date when
the storm center was closest to the county center. Values can range from
-3 to 3 (i.e., at most, you can calculate the total rainfall for the
seven-day window centered on the day when the storm is closest to the
county).}
}
\value{
Returns the same type dataframe as \code{county_rain},
   but with storms listed by community instead of county.
}
\description{
This function takes a dataframe with multi-county communities and returns a
community-level dataframe of "exposed" storms, based on the average distance
between the storm's track and the population-based centers of each county
in the community and the given threshold of rainfall, summed over the days
included in the rainfall measurement.
}
\examples{
communities <- data.frame(commun = c(rep("ny", 6), "no", "new"),
                         fips = c("36005", "36047", "36061",
                                  "36085", "36081", "36119",
                                  "22071", "51700"))
rain_storm_df <- multi_county_rain(communities = communities,
                                   start_year = 1995, end_year = 2005,
                                   rain_limit = 100, dist_limit = 100)

}

