% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_exposure.R
\name{rain_exposure}
\alias{rain_exposure}
\title{Create storm exposure time series files}
\usage{
rain_exposure(counties, start_year, end_year, rain_limit, dist_limit, out_dir)
}
\description{
This function ...
}

